Fix semantic code problems
Removed IAuthenticationPacket.cs
Renamed UaPPacket to UsernameAndPasswordRequest.cs and moved to ./Auth/UsernameAndPassword
Use Encoding.ASCII.GetString() instead of ToString()
Use byte instead of int for Username/Password Length
